reviewbot清理

+++ frameworks/accesstoken/src/hap_policy_parcel.cpp
@@ -31,20 +31,20 @@ bool HapPolicyParcel::Marshalling(Parcel& out) const
     RETURN_IF_FALSE(out.WriteString(this->hapPolicyParameter.domain));
 
     const std::vector<PermissionDef>& permList = this->hapPolicyParameter.permList;
-    int32_t permListSize = static_cast<int32_t>(permList.size());
-    RETURN_IF_FALSE(out.WriteInt32(permListSize));
+    uint32_t permListSize = permList.size();
+    RETURN_IF_FALSE(out.WriteUint32(permListSize));
 
-    for (int i = 0; i < permListSize; i++) {
+    for (uint32_t i = 0; i < permListSize; i++) {
         PermissionDefParcel permDefParcel;
         permDefParcel.permissionDef = permList[i];
         out.WriteParcelable(&permDefParcel);
     }
 
     const std::vector<PermissionStateFull>& permStateList = this->hapPolicyParameter.permStateList;
-    int32_t permStateListSize = static_cast<int32_t>(permStateList.size());
-    RETURN_IF_FALSE(out.WriteInt32(permStateListSize));
+    uint32_t permStateListSize = permStateList.size();
+    RETURN_IF_FALSE(out.WriteUint32(permStateListSize));
 
-    for (int i = 0; i < permStateListSize; i++) {
+    for (uint32_t i = 0; i < permStateListSize; i++) {
         PermissionStateFullParcel permStateParcel;
         permStateParcel.permStatFull = permStateList[i];
         out.WriteParcelable(&permStateParcel);
@@ -66,20 +66,20 @@ HapPolicyParcel* HapPolicyParcel::Unmarshalling(Parcel& in)
 
     hapPolicyParcel->hapPolicyParameter.domain = in.ReadString();
 
-    int permListSize;
-    RELEASE_IF_FALSE(in.ReadInt32(permListSize), hapPolicyParcel);
+    uint32_t permListSize;
+    RELEASE_IF_FALSE(in.ReadUint32(permListSize), hapPolicyParcel);
     RELEASE_IF_FALSE((permListSize <= MAX_PERMLIST_SIZE), hapPolicyParcel);
 
-    for (int i = 0; i < permListSize; i++) {
+    for (uint32_t i = 0; i < permListSize; i++) {
         sptr<PermissionDefParcel> permDefParcel = in.ReadParcelable<PermissionDefParcel>();
         RELEASE_IF_FALSE(permDefParcel != nullptr, hapPolicyParcel);
         hapPolicyParcel->hapPolicyParameter.permList.emplace_back(permDefParcel->permissionDef);
     }
 
-    int permStateListSize;
-    RELEASE_IF_FALSE(in.ReadInt32(permStateListSize), hapPolicyParcel);
+    uint32_t permStateListSize;
+    RELEASE_IF_FALSE(in.ReadUint32(permStateListSize), hapPolicyParcel);
     RELEASE_IF_FALSE((permStateListSize <= MAX_PERMLIST_SIZE), hapPolicyParcel);
-    for (int i = 0; i < permStateListSize; i++) {
+    for (uint32_t i = 0; i < permStateListSize; i++) {
         sptr<PermissionStateFullParcel> permissionStateParcel = in.ReadParcelable<PermissionStateFullParcel>();
         RELEASE_IF_FALSE(permissionStateParcel != nullptr, hapPolicyParcel);
         hapPolicyParcel->hapPolicyParameter.permStateList.emplace_back(permissionStateParcel->permStatFull);
